+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added compatibility with @magento/venia-ui:9.1 (thanks to @yaroslav-qlicks)
+
 ### Fixed
 - Fixed css include on PWA studio v12.1 (thanks to @yaroslav-qlicks)
 - Fixed undefined command: "install-peers" (thanks to @yaroslav-qlicks)